fixing names sscripts

+++ delete_files.py
@@ -15,9 +15,12 @@ def run_paths (_list_paths, input_index=True,  input_images=False):
 print("[INFO] Deleting Files..:")
 for db_paths in _list_paths:	
 	for db_path in db_paths:
-		# get a list of all files in the folder
-		file_list = glob.glob(os.path.join(db_path, "*"))
-		# delete each file in the list
-		for file_path in file_list:
-		    os.remove(file_path)		
-			print(db_sub)
+		if 'raw' in db_sub:
+					None
+		else:		
+			# get a list of all files in the folder
+			file_list = glob.glob(os.path.join(db_path, "*"))
+			# delete each file in the list
+			for file_path in file_list:
+			    os.remove(file_path)		
+				print(db_sub)
